Migrate company `CannotFindMatch` page to React Router (#7467)
* Migrate CannotFindMatch to React Router

* Move CannotFindMatch code to modules

* Remove legacy code

* Move all submission logic to Redux

* Remove legacy code #2
