+++ src/apps/companies/apps/match-company/client/tasks.js
@@ -14,20 +14,6 @@ export const onMatchSubmit = ({ csrfToken, company, dnbCompany }) =>
       return response.data
     })
 
-export const cannotFindMatchSubmit = ({ csrfToken, values, company }) =>
-  axios
-    .post(urls.companies.match.cannotFind(company.id), {
-      _csrf: csrfToken,
-      website: values.website,
-      telephone_number: values.telephoneNumber,
-    })
-    .catch((e) => {
-      return Promise.reject(e.message)
-    })
-    .then((response) => {
-      return response.data
-    })
-
 export async function submitMergeRequest({ company, dnbCompany, csrfToken }) {
   await axios.post(
     `${urls.companies.match.merge(company.id)}?_csrf=${csrfToken}`,
