+++ src/apps/companies/apps/match-company/router.js
@@ -6,15 +6,11 @@ const {
   findDnbCompany,
   renderMatchConfirmation,
   linkCompanies,
-  renderCannotFindMatch,
-  submitNewDnbRecordRequest,
   submitMergeRequest,
 } = require('./controllers')
 
 router.get(urls.companies.match.index.route, renderFindCompanyForm)
 router.post(urls.companies.match.index.route, findDnbCompany)
-router.get(urls.companies.match.cannotFind.route, renderCannotFindMatch)
-router.post(urls.companies.match.cannotFind.route, submitNewDnbRecordRequest)
 router.get(urls.companies.match.confirmation.route, renderMatchConfirmation)
 router.post(urls.companies.match.link.route, linkCompanies)
 router.post(urls.companies.match.merge.route, submitMergeRequest)
