--- src/apps/companies/apps/match-company/views/cannot-find-match.njk
@@ -1,9 +1,0 @@
-{% extends "_layouts/template-no-local-header.njk" %}
-
-{% block body %}
-  {% component 'react-slot', {
-    id: 'cannot-find-match',
-    props: props
-  } %}
-
-{% endblock %}
