+++ src/client/routes.js
@@ -112,6 +112,7 @@ import ReferralDetails from './modules/Companies/Referrals/ReferralDetails'
 import ContactDetails from './modules/Contacts/ContactDetails/ContactDetails'
 import ContactActivity from './modules/Contacts/ContactActivity/ContactActivity'
 import ContactAuditHistory from './modules/Contacts/ContactAuditHistory/ContactAuditHistory'
+import CannotFindMatch from './modules/Companies/MatchCompany/CannotFindMatch'
 
 function Routes() {
   const routes = useRoutes([
@@ -377,6 +378,14 @@ function Routes() {
       element: (
         <ProtectedRoute module={'datahub:companies'}>
           <ReferralDetails />
+        </ProtectedRoute>
+      ),
+    },
+    {
+      path: '/companies/:companyId/match/cannot-find',
+      element: (
+        <ProtectedRoute module={'datahub:companies'}>
+          <CannotFindMatch />
         </ProtectedRoute>
       ),
     },
